Move get_mock_context method to central utils file.
