+++ tests/test_tags.py
@@ -1,19 +1,10 @@
 import unittest
 
-from unittest.mock import patch, MagicMock
-
+from unittest.mock import patch
 
 from datadog_lambda.tags import parse_lambda_tags_from_arn
 
-
-def get_mock_context(
-    invoked_function_arn="arn:aws:lambda:us-east-1:1234597598159:function:swf-hello-test:$Latest",
-    function_version="1",
-):
-    lambda_context = MagicMock()
-    lambda_context.invoked_function_arn = invoked_function_arn
-    lambda_context.function_version = function_version
-    return lambda_context
+from tests.utils import get_mock_context
 
 
 class TestMetricTags(unittest.TestCase):
@@ -23,8 +14,12 @@ class TestMetricTags(unittest.TestCase):
         self.addCleanup(patcher.stop)
 
     def test_parse_lambda_tags_from_arn_latest(self):
+        lambda_context = get_mock_context()
+        lambda_context.invoked_function_arn = (
+            "arn:aws:lambda:us-east-1:1234597598159:function:swf-hello-test:$Latest"
+        )
         self.assertListEqual(
-            parse_lambda_tags_from_arn(get_mock_context()),
+            parse_lambda_tags_from_arn(lambda_context),
             [
                 "region:us-east-1",
                 "account_id:1234597598159",
